Structured CLI tests & utils.
