+++ medcat/cli/system_utils.py
@@ -291,3 +291,15 @@ def dict_diff(dict_1, dict_2):
             
 
     return diff_1, diff_2
+
+def make_orderer():
+    order = {}
+
+    def ordered(f):
+        order[f.__name__] = len(order)
+        return f
+
+    def compare(a, b):
+        return [1, -1][order[a] < order[b]]
+
+    return ordered, compare
